requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.20.38
botocore==1.23.38
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.8
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-etc==1.3.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-js-asset==1.2.2
django-storages==1.12.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.8
validate-email==1.3
whitenoise==5.3.0
zope.interface==5.4.0
